requirements.txt
openai==1.35.3
python-dotenv==1.0.1
scipy==1.13.1
pythainlp==5.0.4
tiktoken==0.7.0
Flask==3.0.3
pandas==2.2.2
openpyxl==3.1.4
google-generativeai
mysql-connector-python==9.0.0
streamlit
gspread
oauth2client
pprintpp